fix ut coverage test=develop
