+++ python/paddle/fluid/tests/unittests/test_dist_mnist_with_program.py
@@ -17,14 +17,29 @@ import unittest
 from test_dist_base import TestDistBase
 
 
-class TestDistMnistWithProgram(TestDistBase):
+class TestDistMnistLocalSGDFleetApi(TestDistBase):
     def _setup_config(self):
         self._sync_mode = True
         self._use_reduce = False
         self._use_reader_alloc = False
         self._nccl2_mode = True
         self._gpu_fleet_api = True
-        self._train_with_program = True
+        self._use_local_sgd = True
+
+    def test_dist_train(self):
+        import paddle.fluid as fluid
+        if fluid.core.is_compiled_with_cuda():
+            self.check_with_place("dist_mnist.py", delta=1e-5)
+
+
+class TestDistMnistGradAllReduceFleetApi(TestDistBase):
+    def _setup_config(self):
+        self._sync_mode = True
+        self._use_reduce = False
+        self._use_reader_alloc = False
+        self._nccl2_mode = True
+        self._gpu_fleet_api = True
+        self._ut4grad_allreduce = True
 
     def test_dist_train(self):
         import paddle.fluid as fluid
